chore: move utils to react directory
